% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_smoking_probabilistic_original.R
\name{markov}
\alias{markov}
\title{Markov smoking model}
\usage{
markov(n.cycles = 100, n.samples = 10000)
}
\arguments{
\item{n.cycles}{Number of cycles}

\item{n.samples}{Number of samples}
}
\value{
Output
}
\description{
Markov smoking model
}
