% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_expanded_probabilistic.R
\name{markov_expanded}
\alias{markov_expanded}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
