% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_expanded_probabilistic_lapply.R
\name{markov_expanded_lapply}
\alias{markov_expanded_lapply}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_lapply()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
