% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_expanded_probabilistic_lapply_rcpp.R
\name{markov_expanded_lapply_rcpp}
\alias{markov_expanded_lapply_rcpp}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_lapply_rcpp()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
