% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/markov_expanded_probabilistic_lapply_vectorised.R
\name{markov_expanded_lapply_vectorise}
\alias{markov_expanded_lapply_vectorise}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_lapply_vectorise()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
