% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/markov_smoking_probabilistic_parallelsmp_furrr.R
\name{markov_expanded_parallisesmp_furrr}
\alias{markov_expanded_parallisesmp_furrr}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_parallisesmp_furrr(
  n.states = 10,
  n.cycles = 100,
  n.samples = 10000
)
}
\arguments{
\item{n.cycles}{Number of cycles}

\item{n.samples}{Number of samples}

\item{n.state}{Number of states}
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
