% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/markov_smoking_probabilistic_parallelsmp_mclapply.R
\name{markov_expanded_parallisesmp_mclapply}
\alias{markov_expanded_parallisesmp_mclapply}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_parallisesmp_mclapply(n.states = 10, n.cycles = 100,
  n.samples = 10000)
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
