% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_expanded_probabilistic_vectorise.R
\name{markov_expanded_vectorise}
\alias{markov_expanded_vectorise}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_vectorise()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
