% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_smoking_probabilistic_vectorisesmp.R
\name{markov_expanded_vectorisesmp}
\alias{markov_expanded_vectorisesmp}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_expanded_vectorisesmp(n.states = 2, n.cycles = 100, n.samples = 10000)
}
\arguments{
\item{n.states}{Number of states}

\item{n.cycles}{Number of cycles}

\item{n.samples}{Number of samples}
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
