% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_smoking_probabilistic_rcpp.R
\name{markov_rcpp}
\alias{markov_rcpp}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_rcpp()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
