% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/markov_smoking_probabilistic_reduced_dimensions.R
\name{markov_reduced_dimensions}
\alias{markov_reduced_dimensions}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_reduced_dimensions()
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
