% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_seabbs.R
\name{markov_seabbs}
\alias{markov_seabbs}
\title{Markov smoking model}
\usage{
markov_seabbs(no_samples = 1000, duration = 100, discount = 1.035)
}
\arguments{
\item{n.cycles}{Numeric, the number of model cycles. Defaults to 10. This is a
combination of time and cycle length (so 5 years with a cycle length of 6 months would give
n.cycles).}

\item{n.samples}{Numeric, the number of samples to take. Defaults to 10000.}
}
\value{
Output
}
\description{
Markov smoking model
}
