% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_smoking_probabilistic_vectorisetx.R
\name{markov_vectorisetx}
\alias{markov_vectorisetx}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_vectorisetx(n.cycles = 100, n.samples = 10000)
}
\arguments{
\item{n.cycles}{Number of cycles}

\item{n.samples}{Number of samples}
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
