% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_smoking_probabilistic_vectorisetxsmp.R
\name{markov_vectorisetxsmp}
\alias{markov_vectorisetxsmp}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_vectorisetxsmp(n.cycles = 100, n.samples = 10000)
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
