% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/markov_smoking_probabilistic_reduced_dimensions.R
\name{markov_reduced_dimensions}
\alias{markov_reduced_dimensions}
\title{Reduced dimensions in markov smoking probabilistic model}
\usage{
markov_reduced_dimensions(n.cycles = 100, n.samples = 10000)
}
\arguments{
\item{n.cycles}{Number of cycles}

\item{n.samples}{Number of samples}
}
\value{
Output
}
\description{
Reduced dimensions in markov smoking probabilistic model
}
