% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markov_seabbs.R
\name{markov_seabbs}
\alias{markov_seabbs}
\title{Markov smoking model}
\usage{
markov_seabbs(no_samples = 10000, duration = 100, discount = 1.035)
}
\arguments{
\item{no_samples}{Numeric, the number of samples to take. Defaults to 10000.}

\item{duration}{Numeric, the number of model cycles. Defaults to 10. This is a
combination of time and cycle length (so 5 years with a cycle length of 6 months would give
n.cycles).}

\item{discount}{Numeric, defaults to 1.035. Discounting to apply to costs and qalys}
}
\value{
Output
}
\description{
Markov smoking model
}
\examples{

markov_seabbs()
}
